requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.3.0=pypi_0
aiofiles=24.1.0=pypi_0
aiogram=3.17.0=pypi_0
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.12=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20241001=pypi_0
annotated-types=0.6.0=py310h06a4308_0
anyio=4.6.2=py310h06a4308_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
bitsandbytes=0.45.2=pypi_0
blake3=1.0.4=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.1.31=hbcca054_0
cachetools=5.5.1=pypi_0
certifi=2025.1.31=py310h06a4308_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
colorful=0.5.6=pypi_0
comm=0.2.2=pyhd8ed1ab_1
compressed-tensors=0.9.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.8=3
cut-cross-entropy=25.1.1=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.12=py310hf71b8c6_0
decorator=5.1.1=pyhd8ed1ab_1
depyf=0.18.0=pypi_0
diffusers=0.32.2=pypi_0
dill=0.3.8=pypi_0
diskcache=5.6.3=pypi_0
distlib=0.3.9=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fastapi=0.115.8=pypi_0
filelock=3.17.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
gguf=0.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310h5eee18b_0
google-api-core=2.24.1=pypi_0
google-auth=2.38.0=pypi_0
googleapis-common-protos=1.67.0=pypi_0
grpcio=1.70.0=pypi_0
h11=0.14.0=py310h06a4308_0
hf-transfer=0.1.9=pypi_0
httpcore=1.0.2=py310h06a4308_0
httptools=0.6.4=pypi_0
httpx=0.27.0=py310h06a4308_0
huggingface-hub=0.28.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
interegular=0.3.3=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.32.0=pyh907856f_0
ipywidgets=8.1.5=py310h06a4308_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=py310h06a4308_0
jiter=0.8.2=pypi_0
jsonlines=4.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab_widgets=3.0.13=py310h06a4308_0
krb5=1.21.3=h143b758_0
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.13.0.11=0
libcurand=10.3.9.55=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lm-format-enforcer=0.10.9=pypi_0
magic-filter=1.0.12=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py310h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mistral-common=1.5.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py310h06a4308_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=py310h06a4308_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-ml-py=12.570.86=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
ollama-python=0.4.7=pyhd8ed1ab_0
openai=1.62.0=pypi_0
opencensus=0.11.4=pypi_0
opencensus-context=0.1.3=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.4.0=h7b32b05_1
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
partial-json-parser=0.2.1.1.post5=pypi_0
peft=0.14.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
pluggy=1.5.0=pypi_0
prometheus-client=0.21.1=pypi_0
prometheus-fastapi-instrumentator=7.0.2=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.2.1=pypi_0
proto-plus=1.26.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.1.1=py310ha75aee5_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
py-spy=0.4.0=pypi_0
pyarrow=19.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pybind11=2.13.6=pypi_0
pycountry=24.6.1=pypi_0
pydantic=2.10.3=py310h06a4308_0
pydantic-core=2.27.1=py310h4aa5aa6_0
pygments=2.19.1=pyhd8ed1ab_0
pytest=8.3.4=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
python_abi=3.10=2_cp310
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2025.1=pypi_0
pyyaml=6.0.2=py310h5eee18b_0
pyzmq=26.2.1=py310h71f11fc_0
ray=2.42.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.20.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.8.0=py310h06a4308_0
shtab=1.7.1=pypi_0
six=1.17.0=pyhd8ed1ab_0
smart-open=7.1.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.0=py310h06a4308_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.45.3=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
tiktoken=0.8.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py310ha75aee5_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.48.3=pypi_0
triton=3.1.0=pypi_0
trl=0.15.0.dev0=pypi_0
typeguard=4.4.1=pypi_0
typing-extensions=4.12.2=py310h06a4308_0
typing_extensions=4.12.2=py310h06a4308_0
tyro=0.9.14=pypi_0
tzdata=2025.1=pypi_0
unsloth=2025.2.5=pypi_0
unsloth-zoo=2025.2.3=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
virtualenv=20.29.2=pypi_0
vllm=0.7.2=pypi_0
wandb=0.19.6=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websockets=14.2=pypi_0
wheel=0.45.1=py310h06a4308_0
widgetsnbextension=4.0.13=py310h06a4308_0
wrapt=1.17.2=pypi_0
xformers=0.0.28.post3=pypi_0
xgrammar=0.1.11=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
